requirements.txt
streamlit
moviepy
openai-whisper
opencv-python
numpy
pillow
imageio-ffmpeg